+++ test/index.js
@@ -1,0 +1,10 @@
+// Sham raf
+if (!global.requestAnimationFrame) {
+    global.requestAnimationFrame = (cb) => {
+        return setTimeout(cb, 0);
+    };
+
+    global.cancelAnimationFrame = (t) => {
+        return clearTimeout(t);
+    };
+}