+++ webpack.config.js
@@ -46,7 +46,7 @@ module.exports = {
             {
                 test: /\.jsx?$/,
                 exclude: /(node_modules|bower_components)/,
-                loaders: ['react-hot', 'babel']
+                loaders: ['babel']
             },
             {
                 test: /\.md/,